shopfloor: split wms utils
