+++ shopfloor_mobile/static/wms/src/components/batch_picking_line_detail.js
@@ -43,7 +43,7 @@ export var batch_picking_line = Vue.component("batch-picking-line-detail", {
   <item-detail-card
     :key="'batch-picking-line-detail-2'"
     :record="line"
-    :options="utils.misc.move_line_product_detail_options()"
+    :options="utils.wms.move_line_product_detail_options()"
     :card_color="utils.colors.color_for(articleScanned ? 'screen_step_done': 'screen_step_todo')"
     />
 
@@ -58,7 +58,7 @@ export var batch_picking_line = Vue.component("batch-picking-line-detail", {
   <v-card class="pa-2" :color="utils.colors.color_for('screen_step_todo')" v-if="showQtyPicker">
     <packaging-qty-picker
       :key="make_component_key(['packaging-qty-picker', line.id])"
-      :options="utils.misc.move_line_qty_picker_options(line)"
+      :options="utils.wms.move_line_qty_picker_options(line)"
       />
   </v-card>
