+++ shopfloor_mobile/static/wms/src/components/detail/detail_package.js
@@ -24,7 +24,7 @@ Vue.component("detail-package", {
                 key_title: "",
                 list_item_options: {
                     fields: this.product_list_fields(),
-                    list_item_klass_maker: this.utils.misc.move_line_color_klass,
+                    list_item_klass_maker: this.utils.wms.move_line_color_klass,
                 },
             };
         },
