+++ shopfloor_mobile/static/wms/src/scenario/location_content_transfer.js
@@ -47,7 +47,7 @@ const LocationContentTransfer = {
                     <item-detail-card
                         :key="make_state_component_key(['detail-move-line-product', rec.id])"
                         :record="rec"
-                        :options="utils.misc.move_line_product_detail_options(rec, {fields: [{path: 'picking.name', label: 'Picking'}]})"
+                        :options="utils.wms.move_line_product_detail_options(rec, {fields: [{path: 'picking.name', label: 'Picking'}]})"
                         :card_color="utils.colors.color_for(state_in(['scan_destination', 'scan_destination_all']) ? 'screen_step_done': 'screen_step_todo')"
                         />
 
@@ -55,7 +55,7 @@ const LocationContentTransfer = {
                             class="pa-2" :color="utils.colors.color_for('screen_step_todo')">
                         <packaging-qty-picker
                             :key="make_state_component_key(['packaging-qty-picker', rec.id])"
-                            :options="utils.misc.move_line_qty_picker_options(rec)"
+                            :options="utils.wms.move_line_qty_picker_options(rec)"
                             />
                     </v-card>
 
@@ -147,7 +147,7 @@ const LocationContentTransfer = {
             return res;
         },
         move_line_detail_list_options: function(move_line) {
-            return this.utils.misc.move_line_product_detail_options(move_line, {
+            return this.utils.wms.move_line_product_detail_options(move_line, {
                 loud_labels: true,
                 fields_blacklist: ["product.qty_available"],
                 fields: [{path: "location_src.name", label: "From"}],
