+++ shopfloor_mobile_base/static/wms/src/components/detail/detail_mixin.js
@@ -45,7 +45,7 @@ export var ItemDetailMixin = {
             return [];
         },
         _render_date(record, field) {
-            return this.utils.misc.format_date_display(_.result(record, field.path));
+            return this.utils.display.format_date_display(_.result(record, field.path));
         },
         has_detail_action(record, field) {
             return _.result(record, field.action_val_path);
